restructured project structure
